filter: calculateRowLength is 32-bit arithmetic, because Java's int is
qpdf/issue-1688a.pdf is 531 bytes and took 435ms where PDFBox takes 0.3ms --
1,449x, and 83% of it was the runtime zeroing memory. It declares

  /DecodeParms << /Predictor 2 /Colors 536870913 /Columns 1 /BitsPerComponent 8 >>

and 536870913 * 8 is 4,294,967,304. Java's int is 32 bits, so that wraps to 8,
bytesPerPixel is 1 and the row is one byte long. Go's int is 64 bits, so the row
was 536,870,913 bytes and decodePredictor allocated two of them: a gigabyte of
zeroed memory from a 531-byte file, and a different answer from the reference.

conventions/java-to-go.md already said which width this wants -- "Java int is
32-bit: use int32 where the width is load-bearing (format fields,
overflow-sensitive arithmetic)" -- and the port had int. The division is inside
the conversion because Java's is inside the int: the whole expression is 32-bit
there, not just the multiply.

Reproducing the wrap is the point rather than a regret. PDFBox's answer for this
file is a one-byte row; a port that computes the arithmetically honest answer
disagrees with the thing it exists to reproduce, and hands an attacker a
gigabyte for 531 bytes. PDFBox's own SECURITY.md puts disproportionate resource
consumption from small attacker-controlled inputs in scope.

435ms to 0.71ms, and the oracle still says 0 of 2 files disagree.

+++ go/pdfbox/filter/predictor.go
@@ -156,8 +156,21 @@ func abs(v int) int {
 // calculateRowLength returns the number of bytes one row occupies, rounded up
 // to a whole byte.
 func calculateRowLength(colors, bitsPerComponent, columns int) int {
-	bitsPerPixel := colors * bitsPerComponent
-	return (columns*bitsPerPixel + 7) / 8
+	// The width is load-bearing, so the arithmetic is done in the width Java
+	// does it in. Java's int is 32 bits and this expression is allowed to wrap;
+	// Go's int is 64 bits and would not, which sounds like an improvement and
+	// is not. qpdf/issue-1688a.pdf is 531 bytes and declares /Colors 536870913:
+	// in 32 bits that multiply wraps to 8 and the row is one byte, which is
+	// what PDFBox answers, and in 64 bits it is 536,870,913 and decodePredictor
+	// allocates two rows of it. A gigabyte of zeroed memory for a 531-byte
+	// file, and a different answer from the reference.
+	//
+	// See conventions/java-to-go.md on int width, JAVA-BUGS.md for the overflow
+	// itself, and TestCalculateRowLengthOverflowsAsJavaDoes.
+	// The division is inside the conversion because Java's is inside the int:
+	// the whole expression is 32-bit there, not just the multiply.
+	bitsPerPixel := int32(colors) * int32(bitsPerComponent)
+	return int((int32(columns)*bitsPerPixel + 7) / 8)
 }
 
 // getBitSeq reads a bit field out of a byte.
